requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
captcha==0.3
flask_mysqldb==0.2.0
Pillow==8.2.0
